backend_FastApi/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-multipart
pyyaml
passlib[bcrypt]
python-jose